requirements.txt
django==2.1.4
graphene-django==2.2.0
django-filter==2.0.0
django-graphql-jwt==0.1.5
postgres==3.0.0
psycopg2-binary==2.8.4
psycopg2==2.8.4